Added option to set the default command key (eg. ! ? - and so on) to make it easier to run the same bot with multiple instances.
Added some better logging to hopefully improve sability later (found one issue so far with is fixed in this commit).

Started to add more functionality to the collectable assembly to built it into a system which can handle unloading if all scripts are unloaded and a way of notifying which scripts havent been unloaded.

Fixed issue where the script was being disposed twice and without a safety check.

Starting to comment the system up a little.